# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateMilestoneRequest:
    properties:
      content:
        example: Приедут с России1
        type: string
      date:
        example: "2026-01-01"
        format: date
        type: string
      process_id:
        example: 1
        type: integer
      title:
        example: Телевидение
        type: string
    type: object
...
  dto.CreateTaskRequest:
    properties:
      end_date:
        example: "2026-02-01"
        format: date
        type: string
      owner_id:
        example: 1
        type: integer
      process_id:
        example: 1
        type: integer
      start_date:
        example: "2026-01-01"
        format: date
        type: string
      title:
        example: Пуско-наладочные работы
        type: string
    type: object
...
  dto.MilestoneResponse:
    properties:
      content:
        example: Приедут с России1
        type: string
      date:
        example: "2026-01-01"
        format: date
        type: string
      id:
        example: 1
        type: integer
      process_id:
        example: 1
        type: integer
      title:
        example: Телевидение
        type: string
    type: object
...
  dto.TaskResponse:
    properties:
      end_date:
        type: string
      id:
        type: integer
      owner_id:
        type: integer
      process_id:
        type: integer
      start_date:
        type: string
      title:
        type: string
    type: object
...
  dto.UpdateMilestoneRequest:
    properties:
      content:
        example: Приедут с России1
        type: string
      date:
        example: "2026-01-01"
        format: date
        type: string
      process_id:
        example: 1
        type: integer
      title:
        example: Телевидение
        type: string
    type: object
...
  dto.UpdateTaskRequest:
    properties:
      end_date:
        example: "2026-02-01"
        format: date
        type: string
      owner_id:
        example: 1
        type: integer
      process_id:
        example: 1
        type: integer
      start_date:
        example: "2026-01-01"
        format: date
        type: string
      title:
        example: Пуско-наладочные работы
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        example: error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: shmonov.mv@gmail.com
    name: Shmonov Matvey
    url: https://t.me/Koshsky
  description: For managing the enterprise's universal resources
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Enterprise Resource Planning
  version: "1.0"
paths:
...
  /milestone:
    get:
      description: List all milestones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MilestoneResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: List milestones
      tags:
      - Milestones
    post:
      consumes:
      - application/json
      description: Create milestone with the input payload
      parameters:
      - description: Milestone data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMilestoneRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MilestoneResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create milestone
      tags:
      - Milestones
  /milestone/{id}:
    delete:
      description: Delete milestone by ID
      parameters:
      - description: Milestone ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete milestone
      tags:
      - Milestones
    get:
      description: Get milestone by id
      parameters:
      - description: Milestone ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MilestoneResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get milestone
      tags:
      - Milestones
    put:
      consumes:
      - application/json
      description: Update milestone by ID
      parameters:
      - description: Milestone ID
        in: path
        name: id
        required: true
        type: integer
      - description: Milestone data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMilestoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.MilestoneResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Update milestone
      tags:
      - Milestones
...
  /task:
    get:
      description: Get a list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TaskResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: List all tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task
      parameters:
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      description: Get a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get a task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update a task
      tags:
      - Tasks
  /timesheet/calendar:
    get:
      description: Get per-day resource availability (capacity/unavailable/available)
        for a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get resource calendar
      tags:
      - TimesheetCalendar
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Provide JWT token in the format: Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"